Removed ndadd
I removed the function 'ndadd' since I don't understand it anymore. It was once a simple function containing nothing but simple operations. But the usage of strange arrow operators
has lead to a complicated jibberish of function. I will completely rewrite the function. It will be the base of the multiplication function using pthreads to split the operations